feat: move sendNewLibraryCardIssuedMessage
